+++ client/src/App.tsx
@@ -11,7 +11,7 @@ import Booking from "pages/Booking"
 import Profile from "pages/Profile"
 import Admin from "pages/Admin"
 import Thanks from "pages/Thanks"
-import AdminBookingsDetailedView from "./pages/AdminBookingsDetailedView"
+import AdminBookingsDetailedView from "pages/AdminBookingsDetailedView"
 import { BrowserRouter as Router, Route, Routes } from "react-router-dom"
 import { ThemeProvider } from "@mui/material"
 import { LocalizationProvider } from "@mui/x-date-pickers"
