+++ client/src/pages/Contact.tsx
@@ -1,5 +1,5 @@
 import React from "react"
-import ContactForm from "components/ContactForm"
+import ContactForm from "components/ContactForm/ContactForm"
 import { Typography, Stack, Container } from "@mui/material"
 
 const Contact = () => {
