+++ client/src/pages/Login.tsx
@@ -1,5 +1,5 @@
 import React from "react"
-import LoginForm from "components/LoginForm"
+import LoginForm from "components/LoginForm/LoginForm"
 import { Typography, Stack, Container, Link } from "@mui/material"
 
 const Login = () => {
