+++ client/src/pages/Profile.tsx
@@ -11,10 +11,10 @@ import {
 } from "firebase/firestore"
 import { Stack, Typography } from "@mui/material"
 import "../styles/Profile.css"
-import ProfileCard from "components/ProfileCard"
-import ProfileCalendarCard from "components/ProfileCalendarCard"
-import ProfileCurrentBookings from "components/ProfileCurrentBookings"
-import ProfileBookingHistory from "components/ProfileBookingHistory"
+import ProfileCard from "components/ProfileCard/ProfileCard"
+import ProfileCalendarCard from "components/ProfileCalendarCard/ProfileCalendarCard"
+import ProfileCurrentBookings from "components/ProfileCurrentBookings/ProfileCurrentBookings"
+import ProfileBookingHistory from "components/ProfileBookingHistory/ProfileBookingHistory"
 import { useAuthenticatedUser } from "../hooks/useAuthenticatedUser"
 
 /**
